fixed warning "unary minus operator applied to unsigned type, result still unsigned" (msvc)
svn path=/trunk/KDE/kdelibs/; revision=793638

+++ khtml/xml/dom2_rangeimpl.cpp
@@ -810,7 +810,7 @@ void RangeImpl::insertNode( NodeImpl *newNode, int &exceptioncode )
             return;
 
         if (m_startContainer == m_endContainer) {
-            endOffsetDelta = -m_startOffset;
+            endOffsetDelta = -(long)m_startOffset;
             setEndContainer(newText);
             // ### what about border cases where start or end are at
             // ### the margins of the text?
